requirements.txt
torch>=2.0.0
python-chess>=1.999
numpy>=1.24.0
tensorboard>=2.13.0

fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0
python-multipart>=0.0.6

pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
tqdm>=4.65.0

pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0

black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
isort>=5.12.0

loguru>=0.7.0

marshmallow>=3.20.0